Add front page portrait randomization script and images
I rewrote the script from scratch around 2009 or so after the previous
version (probably written by Mist/cycholka) was lost in a hard disk
drive crash affecting Wesnoth.org around July/August 2008. Accompanying
the script are the front page portraits in the current rotation as of
April 2014.

Both the script and the portraits have remained unchanged for a few
years.