Initial commit
[[Administrative notice:

This is the Glamdrol theme as it was before the Wesnoth.org asheviere ->
baldras host migration and accompanying MediaWiki upgrade in October
2013. It worked without issues with MediaWiki 1.15.1, requiring no code
modifications.

It is not known whether there are previous versions of Glamdrol at this
time, but archiving them falls outside the scope of this repository.
From this point we should move forwards to build an improved version of
the Glamdrol theme that works with minimal/no issues with modern
versions of MediaWiki such as 1.22.x.]]