Move to GNU GPL v2+ licensing
The license for Glamdrol up to this point is said to be the GNU GPL
without specifying any version in particular. Even in version 1.0 of the
license [1], this should be assumed to grant users the choice between
any version of the license:

[1]: http://www.gnu.org/licenses/old-licenses/gpl-1.0.html

> If the Program does not specify a version number of the license, you
> may choose any version ever published by the Free Software Foundation.

Furthermore, the author of the Glamdrol theme stated his past intention
to license Glamdrol under the same terms as Battle for Wesnoth itself
when asked by email about a specific GNU GPL version:

> I don't remember what version of GPL I released it under, bug [sic] I
> wanted it to be under the same license as Wesnoth itself. So I think
> it should be "GPL version 2 or later".

In general, the GNU GPL version 1 is now obsolete and shouldn't be used,
and there really should be nobody who cares about us dropping that
possibility for future changes.